added test to initialize with int64, reverted ariel's fix.
Ariel's previous fix did allow us to initialize with int64 arrays, but
prevented honoring the copy=False flag for our Epochs implementation. I've
reverted it to the old logic (which still imporperly deals with int64 arrays).

The complete fix may be to no have copy=True as the default flag because we're
usually converting to our internal representation.