FIX: test_viz.py regression with networkx 2.8.8
As seen in nitime github issue #207, fixing the deprecation of
adj_matrix with adjacency_matrix in networkx 3.0 caused a regression
in test_viz.py when running with the slightly older networkx 2.8.8.

Per Chris' suggestion, this patch abandons the approach of using
adjacent matrix methods for the more generic to_numpy_array, which
works unchanged for both networks 2.8.8 and 3.0 onwards.

Thanks: Nilesh Patra and Chris Markiewicz
Signed-off-by: Étienne Mollier <emollier@debian.org>