doc/conf.py: fix after move to nitime/_version.py.
Since introduction of nitime/_version.py through the move to
pyproject.toml, the sphinx documentation configuration script does not
seem to have followed.  Use of the former variables are causing a
variety of errors when trying to rebuild the documentation.  Adjusting
to the new set of variables from nitime/_version.py fixes that.

Signed-off-by: Étienne Mollier <emollier@debian.org>