tridisolve not accesible
I was using your library and I found it a little bit slow. Diving into the code, I discovered that line (884) in "utils.py" that was trying to import the cythonized module _utils but was unreachable. I just added the dot before the import and the import was successful. I don't know if the dot is only in my case, (venv virtual environment, VSCode, amd64 Windows) or if it's a general fix but I feel that I had to propose changes